nofx-ops-panel/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
paramiko==3.4.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
websockets==12.0
aiofiles==23.2.1
python-dotenv==1.0.0
